--- a/cli/flox-rust-sdk/src/providers/upgrade_checks.rs
+++ b/cli/flox-rust-sdk/src/providers/upgrade_checks.rs
@@ -90,7 +90,6 @@ impl UpgradeInformationGuard<Unlocked> {
 
         debug!(
             ?upgrade_information_path,
-            ?info,
             "created unlocked upgrade information guard"
         );